+++ engine/lite/kafka/runtime/src/main/scala/pl/touk/nussknacker/engine/lite/RunnableScenarioInterpreterFactory.scala
@@ -53,7 +53,7 @@ object RunnableScenarioInterpreterFactory extends LazyLogging {
 }
 
 trait RunnableScenarioInterpreter extends AutoCloseable {
-  def routes(): Option[Route]
+  def routes: Option[Route]
   def run(): Future[Unit]
   def status(): TaskStatus
 }
