+++ engine/lite/kafka/runtime/src/main/scala/pl/touk/nussknacker/engine/lite/kafka/HealthCheckRoutesProvider.scala
@@ -19,7 +19,7 @@ class HealthCheckRoutesProvider(system: ActorSystem, scenarioInterpreter: Runnab
 
   private val management = AkkaManagement(system)
 
-  def routes(): Route = management.routes
+  def routes: Route = management.routes
 }
 
 class KafkaRuntimeRunningCheck(system: ActorSystem) extends (() => Future[Boolean]) with LazyLogging {
